fix(parser): Add parsing of function types
